WIP Dynamic Ladder and Menghani & Matani

+++ vec.c
@@ -7,7 +7,6 @@
 
 #include "vec.h"
 
-
 int vec_expand_(char **data, int *length, int *capacity, int memsz) {
   if (*length + 1 > *capacity) {
     void *ptr;
@@ -24,7 +23,7 @@ int vec_expand_(char **data, int *length, int *capacity, int memsz) {
 int vec_reserve_(char **data, int *length, int *capacity, int memsz, int n) {
   (void) length;
   if (n > *capacity) {
-    void *ptr = realloc(*data, n * memsz);
+    void *ptr = realloc(*data, n * (long)memsz);
     if (ptr == NULL) return -1;
     *data = ptr;
     *capacity = n;
